Switch to libunibreak for finding possible breaks
Our own implementation is based on:
http://unicode.org/Public/PROGRAMS/LineBreakSampleCpp/

Which right there it says this is outdated and does not support the
algorithm after Unicode 5.2. Further more the code was even more broken
as there is mismatch between the number of break classes supported by
UCDN abd that outdated code, which was resulting in out of bounds array
access, fun!